Reorganize solution and move AuthorizationResolution folder in Api project to its own project.
