+++ TwitchySharp.Api.AuthorizationResolution.Tests.Unit/TwitchySharp.Api.AuthorizationResolution.Tests.Unit.csproj
@@ -11,14 +11,12 @@
 
   <ItemGroup>
     <PackageReference Include="coverlet.collector" Version="6.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="8.0.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
     <PackageReference Include="xunit" Version="2.5.3" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\TwitchySharp.Api\TwitchySharp.Api.csproj" />
     <ProjectReference Include="..\TwitchySharp.Api.AuthorizationResolution\TwitchySharp.Api.AuthorizationResolution.csproj" />
   </ItemGroup>
