+++ TwitchySharp.Api.AuthorizationResolution/AccessTokens/AppAccessTokens/AppAccessTokenDetails.cs
@@ -1,0 +1,8 @@
+﻿namespace TwitchySharp.Api.AuthorizationResolution;
+
+public record AppAccessTokenDetails
+    : AccessTokenDetails
+{
+    public required ClientIdentity Client { get; init; }
+    public required AppAccessToken AccessToken { get; init; }
+}
