+++ TwitchySharp.Api.AuthorizationResolution/AccessTokens/AppAccessTokens/IResolveAppAccessToken.cs
@@ -1,0 +1,12 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace TwitchySharp.Api.AuthorizationResolution;
+
+/// <summary>
+/// Resolves an <see cref="AppAccessToken"/> for a given <see cref="ClientIdentity"/>.
+/// </summary>
+public interface IResolveAppAccessToken
+{
+    ValueTask<AppAccessToken?> GetToken(ClientIdentity identity, CancellationToken ct = default);
+}
