+++ TwitchySharp.Api.AuthorizationResolution/AccessTokens/ExtensionJsonWebTokens/ExtensionJwtDetails.cs
@@ -1,0 +1,8 @@
+﻿namespace TwitchySharp.Api.AuthorizationResolution;
+
+public record ExtensionJwtDetails
+    : AccessTokenDetails
+{
+    public required ExtensionIdentity Extension { get; init; }
+    public required ExtensionJsonWebToken JsonWebToken { get; init; }
+}
