+++ TwitchySharp.Api.AuthorizationResolution/Clients/SingleClientIdentityResolver.cs
@@ -1,0 +1,15 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace TwitchySharp.Api.AuthorizationResolution;
+
+/// <summary>
+/// A simple <see cref="IResolveClientIdentity"/> implementation that returns the same client identity for all requests.
+/// </summary>
+/// <param name="ClientId">The client identity to return for all requests.</param>
+public record SingleClientIdentityResolver(ClientIdentity ClientId) : IResolveClientIdentity
+{
+    /// <inheritdoc/>
+    public ValueTask<ClientIdentity?> GetClientId(ITwitchRequest request, CancellationToken ct = default)
+        => ValueTask.FromResult<ClientIdentity?>(ClientId);
+}
