+++ TwitchySharp.Api/IRequireAuthorization.cs
