+++ TwitchySharp.Api/TwitchAuthorizationRequestOptions.cs
