+++ TwitchySharp.sln
@@ -41,6 +41,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwitchySharp.Tests", "Twitc
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwitchySharp.Api.Tests.Integration", "TwitchySharp.Api.Tests.Integration\TwitchySharp.Api.Tests.Integration.csproj", "{14721C37-8AC5-42E8-99CF-1F033D6B3980}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwitchySharp.Api.AuthorizationResolution", "TwitchySharp.Api.AuthorizationResolution\TwitchySharp.Api.AuthorizationResolution.csproj", "{C337ECA5-DE14-41D0-8B12-4447816757BC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwitchySharp.Api.AuthorizationResolution.Tests.Unit", "TwitchySharp.Api.AuthorizationResolution.Tests.Unit\TwitchySharp.Api.AuthorizationResolution.Tests.Unit.csproj", "{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -279,6 +283,30 @@ Global
 		{14721C37-8AC5-42E8-99CF-1F033D6B3980}.Release|x64.Build.0 = Release|Any CPU
 		{14721C37-8AC5-42E8-99CF-1F033D6B3980}.Release|x86.ActiveCfg = Release|Any CPU
 		{14721C37-8AC5-42E8-99CF-1F033D6B3980}.Release|x86.Build.0 = Release|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Debug|x64.Build.0 = Debug|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Debug|x86.Build.0 = Debug|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Release|x64.ActiveCfg = Release|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Release|x64.Build.0 = Release|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Release|x86.ActiveCfg = Release|Any CPU
+		{C337ECA5-DE14-41D0-8B12-4447816757BC}.Release|x86.Build.0 = Release|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Debug|x64.Build.0 = Debug|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Debug|x86.Build.0 = Debug|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Release|x64.ActiveCfg = Release|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Release|x64.Build.0 = Release|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Release|x86.ActiveCfg = Release|Any CPU
+		{6A2DE190-B95C-4EF9-8930-9D2FFE2E67EC}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
